feat: refactor ProjectsPage to use ProjectCarousel component and remove unused project components
